refactor: api 테스트 클래스 정리
