+++ src/test/java/com/server/crews/api/ApplicationApiDocuments.java
@@ -1,6 +1,8 @@
 package com.server.crews.api;
 
 import static com.epages.restdocs.apispec.RestAssuredRestDocumentationWrapper.document;
+import static org.springframework.restdocs.payload.PayloadDocumentation.fieldWithPath;
+import static org.springframework.restdocs.payload.PayloadDocumentation.requestFields;
 import static org.springframework.restdocs.request.RequestDocumentation.parameterWithName;
 import static org.springframework.restdocs.request.RequestDocumentation.pathParameters;
 
@@ -13,4 +15,14 @@ public class ApplicationApiDocuments {
         return document(RECRUITMENT_API + "지원서 상세 조회",
                 pathParameters(parameterWithName("application-id").description("지원서 id")));
     }
+
+    public static RestDocumentationFilter GET_APPLICATIONS_200_DOCUMENT() {
+        return document(RECRUITMENT_API + "지원서 목록 조회");
+    }
+
+    public static RestDocumentationFilter EVALUATE_APPLICATIONS_200_DOCUMENT() {
+        return document(RECRUITMENT_API + "지원서 평가",
+                requestFields(fieldWithPath(".passApplicationIds").description("합격 지원서 id 목록")));
+    }
+
 }
