+++ src/test/java/com/server/crews/api/AuthApiTest.java
@@ -1,5 +1,11 @@
 package com.server.crews.api;
 
+import static com.server.crews.api.StatusCodeChecker.checkStatusCode200;
+import static com.server.crews.api.StatusCodeChecker.checkStatusCode401;
+import static com.server.crews.fixture.UserFixture.TEST_EMAIL;
+import static com.server.crews.fixture.UserFixture.TEST_PASSWORD;
+import static org.assertj.core.api.SoftAssertions.assertSoftly;
+
 import com.server.crews.applicant.dto.response.ApplicationsResponse;
 import com.server.crews.auth.dto.request.AdminLoginRequest;
 import com.server.crews.auth.dto.request.ApplicantLoginRequest;
@@ -9,21 +15,14 @@ import com.server.crews.recruitment.dto.response.RecruitmentDetailsResponse;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-
-import java.util.Arrays;
-import java.util.List;
-import java.util.Map;
-
-import static com.server.crews.api.StatusCodeChecker.checkStatusCode200;
-import static com.server.crews.api.StatusCodeChecker.checkStatusCode401;
-import static com.server.crews.fixture.UserFixture.TEST_EMAIL;
-import static com.server.crews.fixture.UserFixture.TEST_PASSWORD;
-import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
 // Todo: 실패 케이스 테스트 추가
 public class AuthApiTest extends ApiTest {
@@ -86,7 +85,8 @@ public class AuthApiTest extends ApiTest {
         AccessTokenResponse adminTokenResponse = signUpAdmin(TEST_EMAIL, TEST_PASSWORD);
         RecruitmentDetailsResponse recruitmentDetailsResponse = createRecruitment(adminTokenResponse.accessToken());
 
-        ApplicantLoginRequest applicantLoginRequest = new ApplicantLoginRequest(recruitmentDetailsResponse.code(), TEST_EMAIL, TEST_PASSWORD);
+        ApplicantLoginRequest applicantLoginRequest = new ApplicantLoginRequest(recruitmentDetailsResponse.code(),
+                TEST_EMAIL, TEST_PASSWORD);
 
         // when
         ExtractableResponse<Response> response = RestAssured.given().log().all()
@@ -117,14 +117,16 @@ public class AuthApiTest extends ApiTest {
         // when
         ExtractableResponse<Response> response = RestAssured.given().log().all()
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .header(HttpHeaders.AUTHORIZATION, AuthorizationExtractor.BEARER_TYPE + adminTokenResponse.accessToken())
+                .header(HttpHeaders.AUTHORIZATION,
+                        AuthorizationExtractor.BEARER_TYPE + adminTokenResponse.accessToken())
                 .when().get("/applications?recruitment-id=" + recruitmentDetailsResponse.id())
                 .then().log().all()
                 .statusCode(HttpStatus.OK.value())
                 .extract();
 
         // then
-        List<ApplicationsResponse> applicationsResponses = Arrays.stream(response.as(ApplicationsResponse[].class)).toList();
+        List<ApplicationsResponse> applicationsResponses = Arrays.stream(response.as(ApplicationsResponse[].class))
+                .toList();
         assertSoftly(softAssertions -> {
             checkStatusCode200(response, softAssertions);
             softAssertions.assertThat(applicationsResponses).hasSize(0);
@@ -137,12 +139,14 @@ public class AuthApiTest extends ApiTest {
         // given
         AccessTokenResponse adminTokenResponse = signUpAdmin(TEST_EMAIL, TEST_PASSWORD);
         RecruitmentDetailsResponse recruitmentDetailsResponse = createRecruitment(adminTokenResponse.accessToken());
-        AccessTokenResponse applicantTokenResponse = signUpApplicant(recruitmentDetailsResponse.code(), TEST_EMAIL, TEST_PASSWORD);
+        AccessTokenResponse applicantTokenResponse = signUpApplicant(recruitmentDetailsResponse.code(), TEST_EMAIL,
+                TEST_PASSWORD);
 
         // when
         ExtractableResponse<Response> response = RestAssured.given().log().all()
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .header(HttpHeaders.AUTHORIZATION, AuthorizationExtractor.BEARER_TYPE + applicantTokenResponse.accessToken())
+                .header(HttpHeaders.AUTHORIZATION,
+                        AuthorizationExtractor.BEARER_TYPE + applicantTokenResponse.accessToken())
                 .when().get("/applications?recruitment-id=" + recruitmentDetailsResponse.id())
                 .then().log().all()
                 .statusCode(HttpStatus.UNAUTHORIZED.value())
