+++ src/test/java/com/server/crews/api/StatusCodeChecker.java
@@ -1,11 +1,11 @@
 package com.server.crews.api;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import org.assertj.core.api.SoftAssertions;
 import org.springframework.http.HttpStatus;
-
-import static org.assertj.core.api.Assertions.assertThat;
 
 public class StatusCodeChecker {
